0.1.0 (#1)
cbuf.js is a drop-in replacement for the
https://github.com/mvi-llc/wasm-cbuf library. It is a full cbuf
implementation (parsing, serialization, and deserialization) written in
TypeScript and a [Nearley](https://nearley.js.org/) grammar.

Parsing in wasm-cbuf suffered from various issues. The C++ code used a
global string allocation pool that leaked memory and would cause
out-of-bounds access when parsing the same .cbuf multiple times, making
the code not reentrant, error handling was done with a mixture of
`printf()` and `exit()` calls, neither of which are a good match for a
WASM library, and the parsing was defined by hand-written C-ish C++ code
that made it difficult to follow what was valid or invalid syntax.